From http://www.jwz.org/xscreensaver/xscreensaver-5.45.tar.gz
-rw-rw-r-- 1 zblaxell zblaxell 27729147 Dec  8 12:55 xscreensaver-5.45.tar.gz
933cd5451bdfc4a2bf15bc49f629a8c8665cae62  xscreensaver-5.45.tar.gz

+++ utils/yarandom.c
@@ -118,9 +118,9 @@ ya_rand_init(unsigned int seed)
 #define ROT(X,N) (((X)<<(N)) | ((X)>>((sizeof(unsigned int)*8)-(N))))
       seed = (999U * (unsigned int) tp.tv_sec);
       seed = ROT (seed, 11);
-      seed += (1001 * tp.tv_usec);
+      seed += (1001 * (unsigned int) tp.tv_usec);
       seed = ROT (seed, 7);
-      seed += (1003 * getpid());
+      seed += (1003 * (unsigned int) getpid());
       seed = ROT (seed, 13);
     }
